requirements.txt
# Core dependencies
PySide6>=6.4.0
python-dotenv>=0.21.0
# Video processing
# Uncomment if implementing actual video processing
moviepy>=1.0.3
yt-dlp>=2023.3.4

# AI and NLP
# Uncomment if implementing actual AI integration
openai>=0.27.0
anthropic>=0.2.8

# Export functionality
# Uncomment if implementing actual export functionality
markdown>=3.4.1
weasyprint>=57.1

# Development tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0

# Added from the code block
requests>=2.28.0